Method call changed

+++ src/Main.java
@@ -2,8 +2,8 @@ public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
 
-        double myIndex = service.calculate(77.54, 1.98);
+        int myIndex = (int) service.calculate(77.54, 1.98);
 
-        System.out.println((int) myIndex);
+        System.out.println(myIndex);
     }
 }
